requirements.txt
Flask>=2.0
Flask-SQLAlchemy
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
whitenoise
passlib~=1.7.4
psycopg2-binary

SQLAlchemy~=1.4.28
marshmallow~=3.14.1
